treewide: Correct unit tests to integration tests
Before this commit, running the standard go command

    go test ./...

would fail because a number of unit tests were in fact integration
tests, typically requiring an etcd server to be running.

This PR adds an integration_test build flag to all such tests and
renames the unit-tests make target to integration-tests and sets this
flag.

With this commit "go test ./..." succeeds and runs a large subset of the
unit tests, completing in less than 15 seconds on a eight core machine
and a fully populated cache. "make integration-tests" in the same
configuration takes about nine minutes as it tests packages serially and
does not use or Go's test result cache.

Unfortunately, the bundling of many tests into gopkg.in/check.v1 suites
means that unit tests in many packages are coupled to integration
tests. This commit makes no attempt to unpick such coupling and so the
set of tests run by go test ./... is smaller than it could potentially
be. Improving this is left for a follow-up commit.
