+++ .travis/build.sh
@@ -5,7 +5,7 @@ set -o errexit
 export CFLAGS="-Werror"
 
 make -j 2
-make unit-tests
+make integration-tests
 
 $HOME/gopath/bin/goveralls -coverprofile=coverage-all.out -service=travis-ci || true
