+++ pkg/endpoint/endpoint_test.go
@@ -1,13 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
-// Copyright 2016-2019 Authors of Cilium
+// Copyright 2016-2021 Authors of Cilium
 
-// +build !privileged_tests
+// +build !privileged_tests,integration_tests
 
 package endpoint
 
 import (
 	"context"
 	"reflect"
+	"strings"
 	"testing"
 	"time"
 
@@ -21,7 +22,7 @@ import (
 	"github.com/cilium/cilium/pkg/identity/cache"
 	ciliumio "github.com/cilium/cilium/pkg/k8s/apis/cilium.io"
 	"github.com/cilium/cilium/pkg/kvstore"
-	pkgLabels "github.com/cilium/cilium/pkg/labels"
+	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/labelsfilter"
 	"github.com/cilium/cilium/pkg/lock"
 	"github.com/cilium/cilium/pkg/maps/ctmap"
@@ -233,23 +234,23 @@ func (s *EndpointSuite) TestEndpointUpdateLabels(c *C) {
 	e := NewEndpointWithState(s, &FakeEndpointProxy{}, &allocator.FakeIdentityAllocator{}, 100, StateWaitingForIdentity)
 
 	// Test that inserting identity labels works
-	rev := e.replaceIdentityLabels(pkgLabels.Map2Labels(map[string]string{"foo": "bar", "zip": "zop"}, "cilium"))
+	rev := e.replaceIdentityLabels(labels.Map2Labels(map[string]string{"foo": "bar", "zip": "zop"}, "cilium"))
 	c.Assert(rev, Not(Equals), 0)
 	c.Assert(string(e.OpLabels.OrchestrationIdentity.SortedList()), Equals, "cilium:foo=bar;cilium:zip=zop;")
 	// Test that nothing changes
-	rev = e.replaceIdentityLabels(pkgLabels.Map2Labels(map[string]string{"foo": "bar", "zip": "zop"}, "cilium"))
+	rev = e.replaceIdentityLabels(labels.Map2Labels(map[string]string{"foo": "bar", "zip": "zop"}, "cilium"))
 	c.Assert(rev, Equals, 0)
 	c.Assert(string(e.OpLabels.OrchestrationIdentity.SortedList()), Equals, "cilium:foo=bar;cilium:zip=zop;")
 	// Remove one label, change the source and value of the other.
-	rev = e.replaceIdentityLabels(pkgLabels.Map2Labels(map[string]string{"foo": "zop"}, "nginx"))
+	rev = e.replaceIdentityLabels(labels.Map2Labels(map[string]string{"foo": "zop"}, "nginx"))
 	c.Assert(rev, Not(Equals), 0)
 	c.Assert(string(e.OpLabels.OrchestrationIdentity.SortedList()), Equals, "nginx:foo=zop;")
 
 	// Test that inserting information labels works
-	e.replaceInformationLabels(pkgLabels.Map2Labels(map[string]string{"foo": "bar", "zip": "zop"}, "cilium"))
+	e.replaceInformationLabels(labels.Map2Labels(map[string]string{"foo": "bar", "zip": "zop"}, "cilium"))
 	c.Assert(string(e.OpLabels.OrchestrationInfo.SortedList()), Equals, "cilium:foo=bar;cilium:zip=zop;")
 	// Remove one label, change the source and value of the other.
-	e.replaceInformationLabels(pkgLabels.Map2Labels(map[string]string{"foo": "zop"}, "nginx"))
+	e.replaceInformationLabels(labels.Map2Labels(map[string]string{"foo": "zop"}, "nginx"))
 	c.Assert(string(e.OpLabels.OrchestrationInfo.SortedList()), Equals, "nginx:foo=zop;")
 }
 
@@ -515,51 +516,51 @@ func (s *EndpointSuite) TestProxyID(c *C) {
 
 func TestEndpoint_GetK8sPodLabels(t *testing.T) {
 	type fields struct {
-		OpLabels pkgLabels.OpLabels
+		OpLabels labels.OpLabels
 	}
 	tests := []struct {
 		name   string
 		fields fields
-		want   pkgLabels.Labels
+		want   labels.Labels
 	}{
 		{
 			name: "has all k8s labels",
 			fields: fields{
-				OpLabels: pkgLabels.OpLabels{
-					OrchestrationInfo: pkgLabels.Map2Labels(map[string]string{"foo": "bar"}, pkgLabels.LabelSourceK8s),
+				OpLabels: labels.OpLabels{
+					OrchestrationInfo: labels.Map2Labels(map[string]string{"foo": "bar"}, labels.LabelSourceK8s),
 				},
 			},
-			want: pkgLabels.Map2Labels(map[string]string{"foo": "bar"}, pkgLabels.LabelSourceK8s),
+			want: labels.Map2Labels(map[string]string{"foo": "bar"}, labels.LabelSourceK8s),
 		},
 		{
 			name: "the namespace labels, service account and namespace should be ignored as they don't belong to pod labels",
 			fields: fields{
-				OpLabels: pkgLabels.OpLabels{
-					OrchestrationInfo: pkgLabels.Map2Labels(map[string]string{
+				OpLabels: labels.OpLabels{
+					OrchestrationInfo: labels.Map2Labels(map[string]string{
 						"foo":                                    "bar",
 						ciliumio.PodNamespaceMetaLabels + ".env": "prod",
 						ciliumio.PolicyLabelServiceAccount:       "default",
 						ciliumio.PodNamespaceLabel:               "default",
-					}, pkgLabels.LabelSourceK8s),
+					}, labels.LabelSourceK8s),
 				},
 			},
-			want: pkgLabels.Map2Labels(map[string]string{"foo": "bar"}, pkgLabels.LabelSourceK8s),
+			want: labels.Map2Labels(map[string]string{"foo": "bar"}, labels.LabelSourceK8s),
 		},
 		{
 			name: "labels with other source than k8s should also be ignored",
 			fields: fields{
-				OpLabels: pkgLabels.OpLabels{
-					OrchestrationInfo: pkgLabels.Map2Labels(map[string]string{
+				OpLabels: labels.OpLabels{
+					OrchestrationInfo: labels.Map2Labels(map[string]string{
 						"foo":                                    "bar",
 						ciliumio.PodNamespaceMetaLabels + ".env": "prod",
-					}, pkgLabels.LabelSourceK8s),
-					OrchestrationIdentity: pkgLabels.Map2Labels(map[string]string{
+					}, labels.LabelSourceK8s),
+					OrchestrationIdentity: labels.Map2Labels(map[string]string{
 						"foo2":                                   "bar",
 						ciliumio.PodNamespaceMetaLabels + ".env": "prod2",
-					}, pkgLabels.LabelSourceAny),
+					}, labels.LabelSourceAny),
 				},
 			},
-			want: pkgLabels.Map2Labels(map[string]string{"foo": "bar"}, pkgLabels.LabelSourceK8s),
+			want: labels.Map2Labels(map[string]string{"foo": "bar"}, labels.LabelSourceK8s),
 		},
 	}
 	for _, tt := range tests {
@@ -709,4 +710,27 @@ func BenchmarkEndpointGetModel(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		e.GetModel()
 	}
+}
+
+// getK8sPodLabels returns all labels that exist in the endpoint and were
+// derived from k8s pod.
+func (e *Endpoint) getK8sPodLabels() labels.Labels {
+	e.unconditionalRLock()
+	defer e.runlock()
+	allLabels := e.OpLabels.AllLabels()
+	if allLabels == nil {
+		return nil
+	}
+
+	allLabelsFromK8s := allLabels.GetFromSource(labels.LabelSourceK8s)
+
+	k8sEPPodLabels := labels.Labels{}
+	for k, v := range allLabelsFromK8s {
+		if !strings.HasPrefix(v.Key, ciliumio.PodNamespaceMetaLabels) &&
+			!strings.HasPrefix(v.Key, ciliumio.PolicyLabelServiceAccount) &&
+			!strings.HasPrefix(v.Key, ciliumio.PodNamespaceLabel) {
+			k8sEPPodLabels[k] = v
+		}
+	}
+	return k8sEPPodLabels
 }
