+++ pkg/endpoint/policy_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-// Copyright 2019 Authors of Cilium
+// Copyright 2019-2021 Authors of Cilium
 
-// +build !privileged_tests
+// +build !privileged_tests,integration_tests
 
 package endpoint
