+++ pkg/identity/cache/allocation_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-// Copyright 2016-2019 Authors of Cilium
+// Copyright 2016-2021 Authors of Cilium
 
-// +build !privileged_tests
+// +build !privileged_tests,integration_tests
 
 package cache
