+++ test/test_test.go
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-// Copyright 2017 Authors of Cilium
+// Copyright 2017-2021 Authors of Cilium
+
+// +build integration_tests
 
 package ciliumTest
